ISSUE #1769: Use eventBus instead of document eventListeners
