+++ public/js/features/chat/chat.js
@@ -1,49 +1,5 @@
-import autoComplete from "../../libraries/autocomplete";
-import { marked } from 'marked';
-
-class ChatMessageRenderer {
-    constructor(messageEl) {
-        this.el = messageEl;
-        this.buffer = '';
-
-        this.md = marked.setOptions({
-            gfm: true,
-            breaks: true,
-        });
-    }
-
-    setText(text) {
-        this.buffer = text;
-    }
-
-    append(chunk) {
-        this.buffer += chunk;
-        this.renderStreaming();
-    }
-
-    renderStreaming() {
-        // VERY lightweight streaming render.
-        const openFences = (this.buffer.match(/```/g) || []).length % 2 === 1;
-
-        let text = this.buffer
-            .replace(/&/g, '&amp;')
-            .replace(/</g, '&lt;')
-            .replace(/>/g, '&gt;');
-
-        if (!openFences) {
-            text = text
-                .replace(/`([^`]+)`/g, '<code>$1</code>')
-                .replace(/\*\*([^*]+)\*\*/g, '<strong>$1</strong>');
-        }
-
-        this.el.innerHTML = text.replace(/\n/g, '<br>');
-    }
-
-    renderFinal() {
-        this.el.dataset.markdownParsed = 'true';
-        this.el.innerHTML = this.md.parse(this.buffer.trim());
-    }
-}
+import autoComplete from "../../../libraries/autocomplete";
+import ChatMessageRenderer from "./message-renderer";
 
 export default class Chat {
     constructor(chatModal) {
@@ -141,7 +97,6 @@ export default class Chat {
     }
 
     bindEvents() {
-        // Form submission
         this.form.addEventListener('submit', e => {
             e.preventDefault();
             const formData = new FormData(this.form);
@@ -149,7 +104,6 @@ export default class Chat {
             this.handleSSE(formData.get('form[message]'));
         });
 
-        // Enter key submits
         this.textInput.addEventListener('keydown', e => {
             if (e.key === 'Enter' && !e.shiftKey) {
                 e.preventDefault();
@@ -157,7 +111,6 @@ export default class Chat {
             }
         });
 
-        // Clear chat
         if (this.clearButton) {
             this.clearButton.addEventListener('click', () => this.clearChat());
         }
