+++ public/js/features/data-table/cluster-renderer.js
@@ -1,5 +1,6 @@
-import Clusterize from '../../libraries/clusterize/clusterize.min';
+import Clusterize from '../../../libraries/clusterize/clusterize.min';
 import {SummableCalculator} from "./summable-calculator";
+import {eventBus, Events} from "../../core/event-bus";
 
 export class ClusterRenderer {
     constructor(wrapper, tbody, scrollElem) {
@@ -23,11 +24,7 @@ export class ClusterRenderer {
                     const resultCount = this.wrapper.querySelector('[data-dataTable-result-count]');
                     if (resultCount) resultCount.innerText = dataRows.filter(r => r.active).length;
 
-                    document.dispatchEvent(new CustomEvent('dataTableClusterWasChanged', {
-                        bubbles: true,
-                        cancelable: false,
-                        detail: {node: this.wrapper}
-                    }));
+                    eventBus.emit(Events.DATA_TABLE_CLUSTER_CHANGED, {node: this.wrapper});
                 }
             }
         });
