+++ public/js/features/data-table/filter-manager.js
